Fixed deb/tmp removal
Build-deb.sh fails if the directory deb/tmp doesn't exist.

+++ deb/build.sh
@@ -14,7 +14,7 @@ fi
 VER=$1
 
 # prepare fresh directories
-rm -rv tmp/
+[ -d tmp ] && rm -rv tmp/
 mkdir -p tmp/DEBIAN/
 
 # changelog
